requirements.txt
antlr4-python3-runtime==4.8
asgiref==3.5.0
Babel==2.9.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
ci-info==0.2.0
click==8.0.3
configobj==5.0.6
configparser==5.2.0
cryptography==36.0.1
decorator==4.4.2
Django==4.0.2
django-ckeditor==6.2.0
django-currencies==0.10.1
Django-Gtts==0.4
django-jazzmin==2.4.9
django-js-asset==2.0.0
django-money==2.1.1
django-multiselectfield==0.1.12
etelemetry==0.3.0
filelock==3.4.2
fitz==0.0.1.dev2
future==0.18.2
gTTS==2.2.3
httplib2==0.20.4
idna==3.3
imageio==2.15.0
imageio-ffmpeg==0.4.5
isodate==0.6.1
jsonfield==2.1.1
lxml==4.7.1
moviepy==1.0.3
networkx==2.6.3
nibabel==3.2.2
nipype==1.7.0
numpy==1.22.2
omegaconf==2.1.1
packaging==21.3
pandas==1.4.1
pathlib==1.0.1
pdfminer.six==20211012
pdfplumber==0.6.0
Pillow==9.0.1
proglog==0.1.9
prov==2.0.0
py-moneyed==1.2
pycparser==2.21
pydot==1.4.2
pydub==0.25.1
PyMuPDF==1.19.5
pyparsing==3.0.7
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2021.3
pyxnat==1.4
PyYAML==6.0
razorpay==1.3.0
rdflib==6.1.1
requests==2.27.1
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
torch==1.10.2
torchaudio==0.10.2
tqdm==4.62.3
traits==6.3.2
typing-extensions==4.1.0
urllib3==1.26.8
Wand==0.6.7
whitenoise==6.0.0
